Move VigilanceName to public API (#57)
* fix oopsie

* apiDump
